Switch to new PRAW search syntax
In March 2018, Reddit killed off a number of filtering features in PRAW, including the timestamp-based
filtering that I previously used here.  Unfortunately, this was really necessary to support batch-based
comment/post processing as I had here, so I had to find a different implementation. This version, due to
being a workaround, has a known bug when a user creates an alert on a phrase that's used in a subreddit
more than 1000 times per hour. This basically never happens - if you were to set up an alert on r/all
for the word "the", you'd *maybe* have a problem - but the bug does exist.