Use post fullname to check if submission is more than an hour old
As opposed to created_utc, which is only convenient for posts and
not comments.

There's still a known bug with this - if a post is marked as the
last post we looked at last hour, and it shows up again in search
results this hour because someone made a comment on it in the last
hour, then Lurkerbot will just stop checking when it sees that post
because, after all, we've already seen the post (but we haven't seen
the comment, which is really what we care about.) Working on trying
to figure out if there's a unified search timestamp I can access
via PRAW in order to fix this.